requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
Flask~=2.3.2
pyyaml~=6.0.1
flask_restful~=0.3.10

